backend/requirements.txt
asgiref==3.3.4
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
distlib==0.3.3
Django==3.2.4
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==21.1
django-taggit==1.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.3.1
gunicorn==20.1.0
idna==3.3
Markdown==3.3.4
Pillow==8.2.0
platformdirs==2.4.0
psycopg2==2.8.6
PyJWT==2.3.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
virtualenv==20.9.0
whitenoise==5.3.0
